[#LR-1390] Overwrite vcr cassettes with new recordings
Some specs needed to be updated because either the actual responses
have changed so the expectations are no longer valid or the specs
depend on data that no longer exists or is now in a different state.

I've removed a couple of specs from `create_shipment_request_spec.rb`
because the nested/shared recordings are quite awkward to work with
and these error conditions are covered by specs in other files.